requirements.txt
asgiref==3.3.1
Django==3.1.7
gunicorn==20.0.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django-cors-headers==3.11.0
djangorestframework==3.12.2
pydub==0.25.1
dj-database-url==0.5.0
psycopg2==2.8.6
audio-metadata==0.11.1
spotdl==3.7.2
daphne==3.0.2
channels==3.0.4
./pytube-master/